pyproject.toml
[project]
name = "wms-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "fastapi>=0.128.0",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
